+++ tf_record_viewer.py
@@ -27,9 +27,9 @@ def decode(example):
         'voxel_obj': tf.FixedLenFeature([], tf.string)
     })
 
-    images = tf.decode_raw(stuff['images'], tf.float64)
+    images = tf.decode_raw(stuff['images'], tf.float32)
     images = tf.reshape(images, (const.N, const.img_resolution, const.img_resolution, 3))
-    depths = tf.decode_raw(stuff['depths'], tf.float64)
+    depths = tf.decode_raw(stuff['depths'], tf.float32)
     depths = tf.reshape(depths, (const.N, const.img_resolution, const.img_resolution, 1))
     # bboxes = tf.decode_raw(stuff['bboxes'], tf.float64)
     # bboxes = tf.reshape(bboxes, (-1, 6))
@@ -40,9 +40,9 @@ def decode(example):
     # anchor_reg = tf.decode_raw(stuff['anchor_reg'], tf.float64)
     # anchor_reg = tf.reshape(anchor_reg, (32, 32, 6))
     # num_obj = tf.decode_raw(stuff['num_obj'], tf.int64)
-    voxel = tf.decode_raw(stuff['voxel'], tf.int64)
+    voxel = tf.decode_raw(stuff['voxel'], tf.float32)
     voxel = tf.reshape(voxel, (128, 128, 128))
-    voxel_obj = tf.decode_raw(stuff['voxel_obj'], tf.int64)
+    voxel_obj = tf.decode_raw(stuff['voxel_obj'], tf.float32)
     voxel_obj = tf.reshape(voxel_obj, (const.max_objects, 128, 128, 128))
     # return images, depths, bboxes, pos_equal_one, neg_equal_one, anchor_reg, num_obj, voxel, voxel_obj
     return images, depths, voxel, voxel_obj
